+++ crates/torii/core/src/processors/store_update_record.rs
