+++ crates/torii/engine/src/constants.rs
@@ -1,10 +1,3 @@
 
 pub(crate) const LOG_TARGET: &str = "torii::engine";
 
-pub const IPFS_URL: &str = "https://ipfs.io/ipfs/";
-pub const IPFS_CLIENT_MAX_RETRY: u8 = 3;
-
-pub const IPFS_CLIENT_URL: &str = "https://ipfs.infura.io:5001";
-pub const IPFS_CLIENT_USERNAME: &str = "2EBrzr7ZASQZKH32sl2xWauXPSA";
-pub const IPFS_CLIENT_PASSWORD: &str = "12290b883db9138a8ae3363b6739d220";
-
