+++ crates/torii/engine/src/lib.rs
@@ -1,5 +1,4 @@
 mod constants;
 mod processors;
-mod utils;
 pub mod engine;
