+++ crates/torii/engine/src/processors/upgrade_event.rs
@@ -8,7 +8,7 @@ use starknet::providers::Provider;
 use tracing::{debug, info};
 
 use super::{EventProcessor, EventProcessorConfig};
-use crate::sql::Sql;
+use torii_sqlite::Sql;
 
 pub(crate) const LOG_TARGET: &str = "torii_core::processors::upgrade_event";
