+++ crates/torii/sqlite/Cargo.toml
@@ -46,3 +46,4 @@ katana-runner.workspace = true
 scarb.workspace = true
 sozo-scarbext.workspace = true
 tempfile.workspace = true
+torii-engine.workspace = true
