+++ crates/torii/sqlite/src/constants.rs
@@ -4,3 +4,10 @@ pub const TOKEN_TRANSFER_TABLE: &str = "token_transfers";
 pub const TOKENS_TABLE: &str = "tokens";
 pub const WORLD_CONTRACT_TYPE: &str = "WORLD";
 pub const SQL_FELT_DELIMITER: &str = "/";
+
+pub const IPFS_URL: &str = "https://ipfs.io/ipfs/";
+pub const IPFS_CLIENT_MAX_RETRY: u8 = 3;
+
+pub const IPFS_CLIENT_URL: &str = "https://ipfs.infura.io:5001";
+pub const IPFS_CLIENT_USERNAME: &str = "2EBrzr7ZASQZKH32sl2xWauXPSA";
+pub const IPFS_CLIENT_PASSWORD: &str = "12290b883db9138a8ae3363b6739d220";
